ADD: install script for arch linux

+++ install.sh
@@ -5,7 +5,11 @@ sudo rm -rf /usr/local/lib/spark
 sudo rm -rf /usr/local/bin/spark
 
 # install prerequisites
-sudo dnf install -y python3 python3-pip
+if [ -f /etc/arch-release ] then
+    sudo pacman -S python python-pip --noconfirm
+elif [ -f /etc/fedora-release ] then
+    sudo dnf install -y python3 python3-pip
+fi
 pip install InquirerPy rich requests
 
 # install spark
